+++ src/app/writer/package-json-writer.ts
@@ -1,14 +1,10 @@
-import pacote from "pacote";
 import * as fs from "fs";
 import * as context from "../context";
 import * as logger from "../../util/logger";
 import { SemVer } from "semver";
 import NPMCliPackageJson from "@npmcli/package-json";
-import { exit } from "process";
 
 export async function writeToFileSystem() {
-  const ndcNodeJsLambdaSdkVersion = await getLatestNdcNodeJsLambdaSdkVersion();
-
   if (packageJsonExists()) {
     logger.info(
       `Updating package.json file found at ${context.getInstance().getOutputDirectory()}`,
@@ -25,7 +21,10 @@ export async function writeToFileSystem() {
   );
 
   const scripts = getScripts(packageJson);
-  const dependencies = getDependencies(packageJson, ndcNodeJsLambdaSdkVersion);
+  const dependencies = getDependencies(
+    packageJson,
+    context.getInstance().getNdcNodeJsLambdaSdkVersion(),
+  );
 
   packageJson.update({
     private: true,
@@ -81,21 +80,4 @@ function getDependencies(
 
 function packageJsonExists(): boolean {
   return fs.existsSync(context.getInstance().getPackageJsonFilePath());
-}
-
-async function getLatestNdcNodeJsLambdaSdkVersion(): Promise<SemVer> {
-  try {
-    const ndcNodeJsLambdaPackageManifest = await pacote.manifest(
-      `@hasura/ndc-lambda-sdk`,
-      {
-        // this is required for custom registry support via .npmrc file (and other npm config options)
-        where: context.getInstance().getUserMountedFilePath(), // Use pacote's where option for npm config resolution (.npmrc file)
-      },
-    );
-
-    return new SemVer(ndcNodeJsLambdaPackageManifest.version);
-  } catch (e) {
-    logger.fatal("Error while downloading dependency @hasura/ndc-lambda-sdk", e);
-    exit(1);
-  }
 }